docs: readme changes (#7042)
* doc: remove note about require and intellisense

This node about using `require()` appears above an example using ES6
imports, which doesn't follow. I suspect this example was updated and
removal of the note was overlooked.

* doc: modernize README.md examples with async/await

Many of the examples use `.then()` and `.catch()` when an `await`
with `try ... catch` would be more concise. This commit does not
change all examples in recognition of the fact that sometimes
`.then()` and `.catch()` are more sensible depending on the context.

---------

Co-authored-by: Jay <jasonsaayman@gmail.com>